+++ migration_wordpress_blog/models/migration_import_wordpress.py
@@ -29,9 +29,19 @@ class MigrationImportWordpressBlog(models.Model):
 
         for blog_json in blog_json_data:
             title = html.unescape(blog_json['title']['rendered'])
+
             slug = blog_json['slug']
+
             content = blog_json['content']['rendered']
             status = blog_json['status']
+
+            feature_media_id = blog_json['featured_media']
+
+            featured_image = False
+            if feature_media_id != "0":
+                response_string = requests.get(self.wordpress_url + "/wp-json/wp/v2/media/" +  str(feature_media_id) )
+                featured_image_json_data = json.loads(response_string.text)
+                featured_image = self.transfer_media( featured_image_json_data )
 
             wraped_content = ""
             wraped_content += "<div class=\"container\">\n"
@@ -52,19 +62,22 @@ class MigrationImportWordpressBlog(models.Model):
             if blog_post:
                 #Update the blog post
                 blog_post.content = transformed_content
+
+                if featured_image:
+                    blog_post.cover_properties = '{"background-image": "url(/web/image/' + str(featured_image.id) + ')", "resize_class": "cover container-fluid cover_full", "background-color": "oe_black", "opacity": 0.2}'
             else:
                 #We also get the Wordpress user and import it if neccassary
                 wordpress_user = self.env['ir.model.data'].xmlid_to_object('wordpress_import.import_user_' + str(blog_json['author']) )
-                if wordpress_user:
-                    _logger.error("TODO implement reimport user code")
-                else:
-                    _logger.error( str(blog_json['author']) )
+                if wordpress_user == False:
                     user_response_string = requests.get(self.wordpress_url + "/wp-json/wp/v2/users/" + str(blog_json['author']) )
                     user_json = json.loads(user_response_string.text)
                     wordpress_user = self.transfer_user(user_json)
 
                 #Create the blog post if it does not exist
                 blog_post = self.env['blog.post'].sudo(wordpress_user.id).create({'author_id': wordpress_user.partner_id.id, 'write_uid': wordpress_user.id, 'blog_id':1, 'name':title, 'content': transformed_content, 'website_published': published})
+
+                if featured_image:
+                    blog_post.cover_properties = '{"background-image": "url(/web/image/' + str(featured_image.id) + ')", "resize_class": "cover container-fluid cover_full", "background-color": "oe_black", "opacity": 0.2}'
 
                 self.env['ir.model.data'].create({'module': "wordpress_import", 'name': external_identifier, 'model': 'blog.post', 'res_id': blog_post.id })
                 self.blog_post_ids = [(4,blog_post.id)]