+++ migration_wordpress/doc/changelog.rst
@@ -1,3 +1,7 @@
+v1.0.1
+======
+* Fix issue with resized images not rendering
+
 v1.0
 ====
 * Port to v11