+++ migration_wordpress_blog/doc/changelog.rst
@@ -1,3 +1,7 @@
+v1.0.1
+======
+* Transfer over cover image( featured image )
+
 v1.0
 ====
 * Port to v11